Attempting to fix the 64-bit builds of Tangra.Core by getting rid of 'long' as from .NET side the P/Invoke mapping is always to 32-bit int regardless of 32/64 bitness.

+++ Tangra.Core/CoreContext.h
@@ -56,7 +56,7 @@ HRESULT InitTangraCore()
 	return S_OK;
 };
 
-HRESULT Set8BitColourChannel(long colourChannel)
+HRESULT Set8BitColourChannel(int colourChannel)
 {
 	s_COLOR_CHANNEL = colourChannel;
