+++ Tangra.Core/Tangra.Pixelmap.cpp
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 
-void SuggestImageDisplayThreasholds(unsigned long* pixels, unsigned char bpp, unsigned int* background, unsigned int* range, unsigned int* histogram)
+void SuggestImageDisplayThreasholds(unsigned int* pixels, unsigned char bpp, unsigned int* background, unsigned int* range, unsigned int* histogram)
 {
 	// TODO:
 }
